fix naming MainActivity.kt move to ui
