+++ app/src/main/java/com/example/myapplication/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.example.myapplication
+package com.example.myapplication.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.myapplication.R
 import com.example.myapplication.ui.home.HomeFragment
 import kotlinx.android.synthetic.main.activity_main.*
